postinst: Do not override hassio.json data if already set
On upgrades, if $DATA_SHARE Is not provided we may end up overwriting
the user configuration, especially in case the postinst script is triggered as
part of dpkg configure because of previous failures (as it may happen in 
recent systemd-journal-remote failures).

To prevent this, read the user value if set and re-use it unless a new one is
provided.