Update README.md to clarify installation instructions
This removes some unnecessary packages from the dependency installation
instructions. They get installed automatically by the package manager
when the main `homeassistant-supervisor.deb` package is installed.

However, the transition from a default Debian networking setup to
NetworkManager and systemd-resolved doesn't work by default.

So this adds installation of the NetworkManager and systemd-resolved
as a separate step. It is a rather critical step and depending on the
network setup it might need manual configuration anyways.