Use stable AppArmor by default
Today, AppArmor profiles for Supervisor are distributed with the channel
as suffix. The profile without a suffix is actually outdated and leads
to Supervisor not being able to initialize udev and raising an Unhealthy
system - Not privileged error (see https://github.com/home-assistant/supervisor/issues/4381).
After the first update the error clears since the AppArmor profile is
updated along with a Supervisor release. However, this makes sure that
Supervised installation start with the proper AppArmor profile from
start.